requirements.txt
fpdf2==2.7.7
pandas==2.1.4
python-dotenv==1.0.0